RabpyWebApp/requirements.txt
django
pillow
djangorestframework
django-rest-swagger
drf_yasg